@@ -2,6 +2,7 @@ package to.itsme.itsmyconfig.tag.impl;
 
 import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
+import to.itsme.itsmyconfig.message.AudienceResolver;
 import to.itsme.itsmyconfig.tag.api.ArgumentsTag;
 import to.itsme.itsmyconfig.util.Scheduler;
 import to.itsme.itsmyconfig.util.Strings;
@@ -31,12 +32,12 @@ public class DelayTag extends ArgumentsTag {
     ) {
         final int delayInTicks = Strings.intOrDefault(arguments[0], 0);
         final String text = arguments[1];
-        final Component component = Utilities.translate(text, player);
-        if (!Component.empty().equals(component)) {
-            Scheduler.runLaterAsync(() -> plugin.adventure().player(player).sendMessage(
-                    component
-            ), delayInTicks);
-        }
+        Scheduler.runLaterAsync(() -> {
+            final Component component = Utilities.translate(text, player);
+            if (!Component.empty().equals(component)) {
+                AudienceResolver.resolve(player).sendMessage(component);
+            }
+        }, delayInTicks);
         return "";
     }
 
